+++ src/neuroconv/datainterfaces/ecephys/phy/phydatainterface.py
@@ -1,6 +1,7 @@
 from pydantic import DirectoryPath, validate_call
 
 from ..basesortingextractorinterface import BaseSortingExtractorInterface
+from ....utils import DeepDict
 
 
 class PhySortingInterface(BaseSortingExtractorInterface):
@@ -42,7 +43,7 @@ class PhySortingInterface(BaseSortingExtractorInterface):
         """
         super().__init__(folder_path=folder_path, exclude_cluster_groups=exclude_cluster_groups, verbose=verbose)
 
-    def get_metadata(self):
+    def get_metadata(self) -> DeepDict:
         metadata = super().get_metadata()
         # See Kilosort save_to_phy() docstring for more info on these fields: https://github.com/MouseLand/Kilosort/blob/main/kilosort/io.py
         # Or see phy documentation: https://github.com/cortex-lab/phy/blob/master/phy/apps/base.py